+++ packages/typespec-client-generator-core/src/interfaces.ts
@@ -377,6 +377,8 @@ export interface SdkDictionaryType extends SdkTypeBase {
 
 export interface SdkNullableType extends SdkTypeBase {
   kind: "nullable";
+  name: string;
+  isGeneratedName: boolean;
   type: SdkType;
   usage: UsageFlags;
   access: AccessFlags;
@@ -530,6 +532,10 @@ export interface SdkEndpointParameter extends SdkModelPropertyTypeBase {
   urlEncode: boolean;
   onClient: true;
   type: SdkEndpointType | SdkUnionType<SdkEndpointType>;
+  /**
+   * @deprecated This property is deprecated. Use `type.templateArguments[x].serializedName` or `type.variantTypes[x].templateArguments[x].serializedName` instead.
+   */
+  serializedName?: string;
 }
 
 export interface SdkCredentialParameter extends SdkModelPropertyTypeBase {