+++ packages/typespec-client-generator-core/src/package.ts
@@ -401,6 +401,8 @@ function getSdkMethodResponse(
   if (nonBodyExists && type) {
     type = {
       kind: "nullable",
+      name: createGeneratedName(context, operation, "NullableResponse"),
+      isGeneratedName: true,
       type: type,
       decorators: [],
       access: "public",